scap next util functions separated from scap_next.h
